+++ CHANGELOG.md
@@ -8,11 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### v3.9.x - [Unreleased]
 
+#### Changed
+
+- Improve Let's Encrypt certificate issuance logging informations
+
 #### Fixed
 
 - Fix cheat.sh install [PR #139](https://github.com/WordOps/WordOps/pull/139)
 - sslutils error when trying to display SSL certificate expiration
-- Fix cheat.sh symbolink link check before creation
+- Fix cheat.sh symbolic link check before creation
 
 ### v3.9.8.9 - 2019-09-03